+++ lib/intervals.dart
@@ -1,3 +1,6 @@
 library intervals;
 
+import 'package:collection/collection.dart';
+
 part 'src/interval.dart';
+part 'src/interval_set.dart';